requirements.txt
numpy==1.16.0
opencv-python==4.0.0.21
pandas==0.23.4
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
six==1.12.0
tqdm==4.29.1
